requirements.txt
fastapi
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
alembic
uvicorn

